python_services/requirements.txt
fastapi
uvicorn[standard]
opencv-python
imgbeddings
psycopg2-binary
numpy
pillow
python-multipart
face-recognition
